// src/types/raw/geocodingRaw.type.ts
export interface GeocodingRawResponse {
    results:           Result[];
    generationtime_ms: number;
}

export interface Result {
    id:           number;
    name:         string;
    latitude:     number;
    longitude:    number;
    elevation:    number;
    feature_code: string;
    country_code: string;
    admin1_id:    number;
    timezone:     string;
    population?:  number;
    country_id:   number;
    country:      string;
    admin1:       string;
}